+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## next ##
 
+* Core: Allow for no (zero-sized) transposition table
 * Engine: Changed defaults and acceptable ranges for options
 * Engine: Exposed QuiescentSearchMaxDepth and PrincipalVariationMaxDepth options
 * Engine: Fixed issue with never using end metric weights